requirements.txt
# AI Counsellor - Backend Dependencies

# Web Framework
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI & External APIs
requests==2.31.0
openai==1.12.0
httpx==0.26.0
google-generativeai==0.8.3

# Email & Utilities
pydantic[email]==2.5.3
pydantic-settings==2.1.0
redis==5.0.1

